requirements.txt
dj-database-url==0.4.2
Django==1.11.4
django-bootstrap3==9.0.0
django-filter==1.0.4
djangorestframework==3.7.0
gunicorn==19.7.1
Markdown==2.6.9
olefile==0.44
Pillow==4.3.0
pytz==2017.2
virtualenv==15.1.0
whitenoise==3.3.1

